Move generic testing utilities into a `testing` package.
This is precursor to moving them into a separate Gradle subproject.
