+++ src/test/java/dev/ionfusion/fusion/CoreTestCase.java
@@ -27,7 +27,7 @@ import com.amazon.ion.IonSystem;
 import com.amazon.ion.IonText;
 import com.amazon.ion.IonValue;
 import com.amazon.ion.system.IonSystemBuilder;
-import dev.ionfusion.fusion.junit.StdioTestCase;
+import dev.ionfusion.testing.StdioTestCase;
 import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.nio.file.Path;
