+++ src/test/java/dev/ionfusion/fusion/FusionRuntimeBuilderTest.java
@@ -6,11 +6,11 @@ package dev.ionfusion.fusion;
 import static dev.ionfusion.fusion.FusionRuntimeBuilder.standard;
 import static dev.ionfusion.fusion.FusionString.unsafeStringToJavaString;
 import static dev.ionfusion.fusion.TestSetup.fusionBootstrapDirectory;
-import static dev.ionfusion.fusion.junit.Reflect.assertEqualProperties;
-import static dev.ionfusion.fusion.junit.Reflect.getterFor;
-import static dev.ionfusion.fusion.junit.Reflect.invoke;
-import static dev.ionfusion.fusion.junit.Reflect.setterFor;
-import static dev.ionfusion.fusion.junit.Reflect.witherFor;
+import static dev.ionfusion.testing.Reflect.assertEqualProperties;
+import static dev.ionfusion.testing.Reflect.getterFor;
+import static dev.ionfusion.testing.Reflect.invoke;
+import static dev.ionfusion.testing.Reflect.setterFor;
+import static dev.ionfusion.testing.Reflect.witherFor;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.instanceOf;
 import static org.junit.jupiter.api.Assertions.assertEquals;
